+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes will be documented in this file
 
+## 1.2.0
+
+ - Add support for Secure Fields and wallets with `secured_card_data field`
+ - Add support for Rest Client 2.0 (thanks @brocktimus)
+
 ## 1.1.0
 
  - Add settlement search support